+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project will be documented in this file.
 - Swapped default single and consume keybinds
 - Clear album art and song in the header when the playback stops
 - Refactored confirm modal into a generic one
+- Refactored rename playlist and save queue modal into a generic modal with single input
 
 ### Fixed